+++ src/term.rs
@@ -8,7 +8,7 @@ use std::{
 use anyhow::{bail, format_err, Error, Result};
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Clone, Copy, PartialEq, Eq)]
 #[repr(u8)]
 pub(crate) enum Coloring {
     Auto = 0,
@@ -20,6 +20,14 @@ impl Coloring {
     const AUTO: u8 = Coloring::Auto as _;
     const ALWAYS: u8 = Coloring::Always as _;
     const NEVER: u8 = Coloring::Never as _;
+
+    pub(crate) fn as_str(self) -> &'static str {
+        match self {
+            Self::Auto => "auto",
+            Self::Always => "always",
+            Self::Never => "never",
+        }
+    }
 }
 
 impl FromStr for Coloring {
@@ -36,9 +44,9 @@ impl FromStr for Coloring {
 }
 
 static COLORING: AtomicU8 = AtomicU8::new(Coloring::AUTO);
-pub(crate) fn set_coloring(color: Option<&str>) -> Result<()> {
+pub(crate) fn set_coloring(color: Option<Coloring>) -> Result<()> {
     let mut coloring = match color {
-        Some(color) => color.parse().map_err(|e| format_err!("argument for --color {}", e))?,
+        Some(color) => color,
         // https://doc.rust-lang.org/nightly/cargo/reference/config.html#termcolor
         None => match env::var_os("CARGO_TERM_COLOR") {
             Some(color) => color
@@ -51,7 +59,6 @@ pub(crate) fn set_coloring(color: Option<&str>) -> Result<()> {
     if coloring == Coloring::Auto && !atty::is(atty::Stream::Stderr) {
         coloring = Coloring::Never;
     }
-    // Relaxed is fine because only the argument parsing step updates this value.
     COLORING.store(coloring as _, Ordering::Relaxed);
     Ok(())
 }
@@ -64,33 +71,30 @@ fn coloring() -> ColorChoice {
     }
 }
 
-static HAS_ERROR: AtomicBool = AtomicBool::new(false);
-pub(crate) fn set_error() {
-    HAS_ERROR.store(true, Ordering::SeqCst)
+macro_rules! global_flag {
+    ($name:ident: $value:ty = $ty:ident::new($($default:expr)?)) => {
+        pub(crate) mod $name {
+            use super::*;
+            pub(super) static VALUE: $ty = $ty::new($($default)?);
+            pub(crate) fn set(value: $value) {
+                VALUE.store(value, Ordering::Relaxed);
+            }
+        }
+        pub(crate) fn $name() -> $value {
+            $name::VALUE.load(Ordering::Relaxed)
+        }
+    };
 }
-pub(crate) fn has_error() -> bool {
-    HAS_ERROR.load(Ordering::SeqCst)
-}
+global_flag!(verbose: bool = AtomicBool::new(false));
+global_flag!(error: bool = AtomicBool::new(false));
+// global_flag!(warn: bool = AtomicBool::new(false));
 
-static VERBOSE: AtomicBool = AtomicBool::new(false);
-pub(crate) fn set_verbose() {
-    // Relaxed is fine because only the argument parsing step updates this value.
-    VERBOSE.store(true, Ordering::Relaxed)
-}
-pub(crate) fn verbose() -> bool {
-    VERBOSE.load(Ordering::Relaxed)
-}
-
-pub(crate) fn print_status(status: &str, color: Option<Color>, justified: bool) -> StandardStream {
+pub(crate) fn print_status(status: &str, color: Option<Color>) -> StandardStream {
     let mut stream = StandardStream::stderr(coloring());
     let _ = stream.set_color(ColorSpec::new().set_bold(true).set_fg(color));
-    if justified {
-        let _ = write!(stream, "{:>12}", status);
-    } else {
-        let _ = write!(stream, "{}", status);
-        let _ = stream.set_color(ColorSpec::new().set_bold(true));
-        let _ = write!(stream, ":");
-    }
+    let _ = write!(stream, "{}", status);
+    let _ = stream.set_color(ColorSpec::new().set_bold(true));
+    let _ = write!(stream, ":");
     let _ = stream.reset();
     let _ = write!(stream, " ");
     stream
@@ -99,8 +103,8 @@ pub(crate) fn print_status(status: &str, color: Option<Color>, justified: bool) 
 macro_rules! error {
     ($($msg:expr),* $(,)?) => {{
         use std::io::Write;
-        crate::term::set_error();
-        let mut stream = crate::term::print_status("error", Some(termcolor::Color::Red), false);
+        crate::term::error::set(true);
+        let mut stream = crate::term::print_status("error", Some(termcolor::Color::Red));
         let _ = writeln!(stream, $($msg),*);
     }};
 }
@@ -108,7 +112,8 @@ macro_rules! error {
 // macro_rules! warn {
 //     ($($msg:expr),* $(,)?) => {{
 //         use std::io::Write;
-//         let mut stream = crate::term::print_status("warning", Some(termcolor::Color::Yellow), false);
+//         crate::term::warn::set(true);
+//         let mut stream = crate::term::print_status("warning", Some(termcolor::Color::Yellow));
 //         let _ = writeln!(stream, $($msg),*);
 //     }};
 // }
@@ -116,15 +121,7 @@ macro_rules! error {
 macro_rules! info {
     ($($msg:expr),* $(,)?) => {{
         use std::io::Write;
-        let mut stream = crate::term::print_status("info", None, false);
+        let mut stream = crate::term::print_status("info", None);
         let _ = writeln!(stream, $($msg),*);
     }};
 }
-
-// macro_rules! status {
-//     ($status:expr, $($msg:expr),* $(,)?) => {{
-//         use std::io::Write;
-//         let mut stream = crate::term::print_status($status, Some(termcolor::Color::Cyan), true);
-//         let _ = writeln!(stream, $($msg),*);
-//     }};
-// }